Add WIB2 frame code
Adds WIB2Frame which interprets a WIB2 frame in an artdaq-like
"overlay" style, and a set of standalone functions which pack and
unpack whole WIB2 frames. The standalone functions are taken from the
dune-wib-firmware respository, written by Ben Land, and are used in
the check_frame_vs_unpack test to check the output of the ADC
interpretation in WIB2Frame.